move services bo burdel sie robi and add item when no list or closed
